src/ConsumptionTaxLawFill_in_the_BlankQuestions/scraper.py
import requests
from bs4 import BeautifulSoup

class Scraper:
    def __init__(self, url):
        self.url = url

    def scrape(self):
        response = requests.get(self.url)
        soup = BeautifulSoup(response.content, 'html.parser')

        # ここでページから必要なデータを抽出します。
        # 例えば、特定のタグやクラスを持つ要素を探す等。
        # 以下は仮のコードです。
        questions = soup.find_all(...)
        answers = soup.find_all(...)

        # 問題と答えを辞書形式で返します。
        questions_answers = {}
        for question, answer in zip(questions, answers):
            questions_answers[question.text] = answer.text
        return questions_answers

